requirements.txt
streamlit==1.42.0
streamlit-authenticator==0.4.1
streamlit-webrtc==0.47.9
av==12.3.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.30
numpy==1.26.4
scipy==1.15.2
scikit-learn==1.6.1
google-generativeai==0.8.4
plotly==6.0.0
pandas==2.2.3
fpdf2==2.8.2
Pillow==11.1.0
sqlalchemy==2.0.38
pyyaml==6.0.2
bcrypt==4.2.1
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn==0.34.0
